refactor(ingestion): move helpers and drop job listener
