+++ api/src/app/agent_tasks/shared/__init__.py
@@ -1,5 +1,11 @@
-"""
-Package: agent_tasks.shared
-Shared schemas and utilities across agent layers.
-"""
-__all__ = []
+"""Shared helpers for agent tasks."""
+
+from .splitter import hash_block, normalise_newlines, parse_blocks
+from .intent import extract_intent
+
+__all__ = [
+    "extract_intent",
+    "hash_block",
+    "normalise_newlines",
+    "parse_blocks",
+]
