--- api/src/app/ingestion/__init__.py
@@ -1,5 +1,0 @@
-"""Ingestion sub-package.
-
-Holds low-level file/text parsing utilities. Later tasks will add
-image OCR, PDF page splitters, etc.
-"""
