+++ api/src/app/routes/dump.py
@@ -7,8 +7,8 @@ import uuid
 
 from fastapi import APIRouter, File, Form, HTTPException, UploadFile, status
 
-from app.ingestion.intent import extract_intent
-from app.ingestion.splitter import parse_blocks
+from app.agent_tasks.shared.intent import extract_intent
+from app.agent_tasks.shared.splitter import parse_blocks
 
 from ..utils.supabase_client import supabase_client as supabase
