+++ codex/PRD/architecture/ls_tree_api_src.md
@@ -60,7 +60,6 @@
     │   └── test_dump_endpoint.py
     └── ingestion
         ├── __init__.py
-        ├── test_job_listener.py
         └── test_splitter_newlines.py
 
-19 directories, 44 files
+19 directories, 43 files
