+++ docs/env.md
@@ -1,5 +1,5 @@
 # Environment Variables
 
 The project uses several environment variables. Supabase-related keys are
-documented in [env_supabase_reference.md](env_supabase_reference.md). Refer to
-that file for canonical names and deployment rules.
+documented in the root [`.env.template`](../.env.template). Refer to that file
+for canonical names and deployment rules.
